% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit conversion}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Celsius temperatures to degrees Fahrenheit
}
\examples{
temp3 <- C_to_F(-273);
temp4 <- C_to_F( c(37, 38, 0, 100) );
}
