% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C_to_K}
\alias{C_to_K}
\title{Celsius to Kelvin conversion}
\usage{
C_to_K(C_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Kelvin
}
\description{
Convert degrees Celsius temperatures to degrees Kelvin
}
\examples{
temp5 <- C_to_K(-273);
temp6 <- C_to_K( c(37, 38, 0, 100) );
}
