% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin to Fahrenheit conversion}
\usage{
K_to_F(K_temp)
}
\arguments{
\item{K_temp}{The temperature in degrees Kelvin}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Kelvin temperatures to degrees Fahrenheit
}
\examples{
temp9 <- K_to_F(0);
temp10 <- K_to_F( c(310, 311, 273, 373) );
}
